refactor: use core/noncopyable over boost/noncopyable
The later is deprecated.
