+++ include/boost/signals2/connection.hpp
@@ -15,9 +15,9 @@
 #define BOOST_SIGNALS2_CONNECTION_HPP
 
 #include <boost/config.hpp>
+#include <boost/core/noncopyable.hpp>
 #include <boost/function.hpp>
 #include <boost/mpl/bool.hpp>
-#include <boost/noncopyable.hpp>
 #include <boost/shared_ptr.hpp>
 #include <boost/signals2/detail/auto_buffer.hpp>
 #include <boost/signals2/detail/null_output_iterator.hpp>
