+++ include/boost/signals2/detail/unique_lock.hpp
@@ -12,7 +12,7 @@
 #ifndef BOOST_SIGNALS2_UNIQUE_LOCK_HPP
 #define BOOST_SIGNALS2_UNIQUE_LOCK_HPP
 
-#include <boost/noncopyable.hpp>
+#include <boost/core/noncopyable.hpp>
 
 namespace boost
 {
